Make entity rule managers
