+++ src/main/kotlin/me/iris/noxesiumapi/entityrules/BeamManager.kt
@@ -10,33 +10,14 @@ import org.bukkit.entity.EnderCrystal
 import java.awt.Color
 import java.util.*
 
-public class BeamColor {
+public class BeamManager(private var entity: Entity) {
     // Only works for minecraft:guardian, minecraft:elder_guardian and minecraft:end_crystal
     private val entityManager = NoxesiumAPI.entityRuleManager
-
-    // Preset colors for Skript
-    public companion object {
-        public val RED: Optional<Color> = Optional.of(Color.RED)
-        public val ORANGE: Optional<Color> = Optional.of(Color(255, 157, 0))
-        public val YELLOW: Optional<Color> = Optional.of(Color.YELLOW)
-        public val LIME: Optional<Color> = Optional.of(Color.GREEN)
-        public val GREEN: Optional<Color> = Optional.of(Color(0, 100, 0))
-        public val CYAN: Optional<Color> = Optional.of(Color(0, 170, 170))
-        public val AQUA: Optional<Color> = Optional.of(Color.CYAN)
-        public val BLUE: Optional<Color> = Optional.of(Color.BLUE)
-        public val PURPLE: Optional<Color> = Optional.of(Color(100, 0, 200))
-        public val PINK: Optional<Color> = Optional.of(Color.MAGENTA)
-        public val WHITE: Optional<Color> = Optional.of(Color.WHITE)
-        public val BLACK: Optional<Color> = Optional.of(Color.BLACK)
-        public val GRAY: Optional<Color> = Optional.of(Color.GRAY)
-        public val LIGHT_GRAY: Optional<Color> = Optional.of(Color.LIGHT_GRAY)
-        public val DARK_GRAY: Optional<Color> = Optional.of(Color.DARK_GRAY)
-    }
 
     /**
      * Set the beam color for a [Guardian] or [EnderCrystal] (also disables the bubbles on guardian beams)
      */
-    public fun setBeamColor(entity: Entity, color: Optional<Color>) {
+    public fun setBeamColor(color: Optional<Color>) {
         if (entity.type != EntityType.GUARDIAN && entity.type != EntityType.END_CRYSTAL) return
         var rule: RemoteServerRule<Any>? = entityManager.getEntityRule(entity, EntityRuleIndices.BEAM_COLOR)
         rule!!.setValue(color)
@@ -50,7 +31,7 @@ public class BeamColor {
     /**
      * Set the beam fade/gradient for a [Guardian] or [EnderCrystal]
      */
-    public fun setBeamFade(entity: Entity, color: Optional<Color>) {
+    public fun setBeamFade(color: Optional<Color>) {
         if (entity.type != EntityType.GUARDIAN && entity.type != EntityType.END_CRYSTAL) return
         val rule: RemoteServerRule<Any>? = entityManager.getEntityRule(entity, EntityRuleIndices.BEAM_COLOR_FADE)
         rule!!.setValue(color)
@@ -59,7 +40,7 @@ public class BeamColor {
     /**
      * Reset the beam color for a [Guardian] or [EnderCrystal]
      */
-    public fun resetBeamColor(entity: Entity) {
+    public fun resetBeamColor() {
         if (entity.type != EntityType.GUARDIAN && entity.type != EntityType.END_CRYSTAL) return
         val rule: RemoteServerRule<Any>? = entityManager.getEntityRule(entity, EntityRuleIndices.BEAM_COLOR)
         rule!!.reset()
@@ -68,7 +49,7 @@ public class BeamColor {
     /**
      * Reset the beam fade/gradient for a [Guardian] or [EnderCrystal]
      */
-    public fun resetBeamFade(entity: Entity) {
+    public fun resetBeamFade() {
         if (entity.type != EntityType.GUARDIAN && entity.type != EntityType.END_CRYSTAL) return
         val rule: RemoteServerRule<Any>? = entityManager.getEntityRule(entity, EntityRuleIndices.BEAM_COLOR_FADE)
         rule!!.reset()
