+++ src/main/kotlin/me/iris/noxesiumapi/util/Colors.kt
@@ -1,0 +1,27 @@
+package me.iris.noxesiumapi.util
+
+import java.awt.Color
+import java.util.*
+
+public class Colors {
+
+    // Preset colors for Skript
+    public companion object {
+        public val RED: Optional<Color> = Optional.of(Color.RED)
+        public val ORANGE: Optional<Color> = Optional.of(Color(255, 157, 0))
+        public val YELLOW: Optional<Color> = Optional.of(Color.YELLOW)
+        public val LIME: Optional<Color> = Optional.of(Color.GREEN)
+        public val GREEN: Optional<Color> = Optional.of(Color(0, 100, 0))
+        public val CYAN: Optional<Color> = Optional.of(Color(0, 170, 170))
+        public val AQUA: Optional<Color> = Optional.of(Color.CYAN)
+        public val BLUE: Optional<Color> = Optional.of(Color.BLUE)
+        public val PURPLE: Optional<Color> = Optional.of(Color(100, 0, 200))
+        public val PINK: Optional<Color> = Optional.of(Color.MAGENTA)
+        public val WHITE: Optional<Color> = Optional.of(Color.WHITE)
+        public val BLACK: Optional<Color> = Optional.of(Color.BLACK)
+        public val GRAY: Optional<Color> = Optional.of(Color.GRAY)
+        public val LIGHT_GRAY: Optional<Color> = Optional.of(Color.LIGHT_GRAY)
+        public val DARK_GRAY: Optional<Color> = Optional.of(Color.DARK_GRAY)
+    }
+
+}
